+++ src/Krystal/Filesystem/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.3
 ---
 
+ * Added `hasExtension()` in `Krystal\Filesystem\FileManager`
  * Changed all methods in `FileManager` to static
  * Fixed `FileManager::rmdir()` non-ability to remove hidden files